Stats country lookup in json update
That a country’s entire data set exists in the json is not an
invariant. This update sets missing attributes to 0. I have also
updated the code that generated the set of arbitrary countries for
chart 3 to fix a bug that could potentially cause the entire set not to
be populated.