Stats and Map updates
* Added shadow separators to the stats page
* Modified the way glows are shown - We take into account now the
number of map_geos we have and set the number of simultaneous glows to
the min of that and our constant of 400. We do not use just the
constant so that in the case that it is erroneously set too high, we
don't end up showing all the map_geos in the first 10 seconds.