Convert title text to title case for EN only.
Adding "The" to the title broke l10n. However,
we did have the string "The Web we want" fully
translated. It was decided to convert the EN string
back to title case automatically to avoid breaking
l10n. I added a filter that allows us to do this
for any languages we choose.